Refactor - ready for testing
