+++ src/neural_net/NeuralNet.java
@@ -5,10 +5,10 @@
  */
 package neural_net;
 
-import common.Sigmoid;
-import common.InitializerIntf;
-import common.InitializerConstant;
 import common.InitializerRandom;
+import common.InitializerIntf;
+import common.RangeTransformIntf;
+import common.Sigmoid;
 import java.util.ArrayList;
 
 /**
@@ -21,75 +21,55 @@ public class NeuralNet implements NeuralNetIntf {
 
 //<editor-fold defaultstate="collapsed" desc="Constructor and Factory Methods">
     {
-        perceptionLayer = new NeuralLayer();
-        hiddenLayer = new NeuralLayer();
-        outputLayer = new NeuralLayer();
-
+        layers = new ArrayList<>();
         learningRate = DEFAULT_LEARNING_RATE;
     }
 
     public NeuralNet() {}
-    
-    public static NeuralNetIntf neuralNetFactory(int perceptionLayerCount, int hiddenLayerCount, int outputLayerCount) {
-        NeuralNetIntf net = new NeuralNet();
+
+    public NeuralNet(int[] layerNodeCounts, InitializerIntf neuronBiasInitializer, NeuralLayerFactoryIntf layerFactory) {
+        // create all the layers 
+        for (int nodeCount : layerNodeCounts){
+            layers.add(layerFactory.getLayer(nodeCount, neuronBiasInitializer));
+        }
         
-        InitializerIntf initializer = new InitializerRandom();
-        //set up the layers
-        net.setPerceptionLayer(new NeuralLayer(perceptionLayerCount, new InitializerConstant(0.0)));
-        net.setHiddenLayer(new NeuralLayer(hiddenLayerCount, initializer));
-        net.setOutputLayer(new NeuralLayer(outputLayerCount, initializer));
-                
-        //connect layers
-        net.getHiddenLayer().connectProviderLayer(net.getPerceptionLayer(), initializer);
-        net.getOutputLayer().connectProviderLayer(net.getHiddenLayer(), initializer);
-        
-        return net;
+        // connect the layers; note that the "input" (or perception) layer is
+        // not connected to other layers, as it will receive the input data
+        for (int i = 1; i < layers.size(); i++){
+            layers.get(i).connectProviderLayer(layers.get(i-1), neuronBiasInitializer);
+        }
     }
     
+    public static NeuralNetIntf getSimpleNeuralNet(int inputLayerCount, int transformLayerCount, int outputLayerCount){
+        int[] layerNodeCounts = {inputLayerCount};
+        return new NeuralNet(layerNodeCounts, new InitializerRandom(), new NeuralLayer());
+    }    
 //</editor-fold>
 
 //<editor-fold defaultstate="collapsed" desc="Properties">
+    private ArrayList<NeuralLayerIntf> layers;
     public static final double DEFAULT_LEARNING_RATE = 0.5;
-
     private double learningRate;
-    private NeuralLayerIntf perceptionLayer, hiddenLayer, outputLayer;
-
+    private static final int INPUT_LAYER_IDX = 0;
+    
     @Override
-    public void setPerceptionLayer(NeuralLayerIntf perceptionLayer) {
-        this.perceptionLayer = perceptionLayer;
+    public ArrayList<NeuralLayerIntf> getLayers() {
+        return layers;
     }
 
     @Override
-    public NeuralLayerIntf getPerceptionLayer() {
-        return perceptionLayer;
+    public NeuralLayerIntf getLayer(int index){
+        return layers.get(index);
     }
 
-    
-
-    
     @Override
     public NeuralLayerIntf getInputLayer() {
-        return getPerceptionLayer();
-    }
-
-    @Override
-    public void setHiddenLayer(NeuralLayerIntf hiddenLayer) {
-        this.hiddenLayer = hiddenLayer;
-    }
-
-    @Override
-    public NeuralLayerIntf getHiddenLayer() {
-        return hiddenLayer;
-    }
-
-    @Override
-    public void setOutputLayer(NeuralLayerIntf outputLayer) {
-        this.outputLayer = outputLayer;
+        return layers.get(INPUT_LAYER_IDX);
     }
 
     @Override
     public NeuralLayerIntf getOutputLayer() {
-        return outputLayer;
+        return layers.get(layers.size() - 1);
     }
 
     @Override
@@ -103,87 +83,76 @@ public class NeuralNet implements NeuralNetIntf {
     }
 //</editor-fold>
 
+//<editor-fold defaultstate="collapsed" desc="Methods">
+
 //<editor-fold defaultstate="collapsed" desc="NeuralNetIntf Methods">
     @Override
     public void pulse() {
-        hiddenLayer.pulse();
-        outputLayer.pulse();
+        for (int i = 1; i < layers.size(); i++){
+            getLayer(i).pulse();
+        }
     }
 
     @Override
     public void applyLearning() {
-        hiddenLayer.applyLearning(learningRate);
-        outputLayer.applyLearning(learningRate);
+        for (int i = 1; i < layers.size(); i++){
+            getLayer(i).applyLearning(learningRate);
+        }
     }
     
     @Override
     public void initializeLearning() {
-        hiddenLayer.initializeLearning();
-        outputLayer.initializeLearning();
+        for (int i = 1; i < layers.size(); i++){
+            getLayer(i).initializeLearning();
+        }
     }   
-//</editor-fold>
 
+    @Override
     public void train(double[][] inputs, double[][]expected, int iterationLimit){
         for (int iteration = 0; iteration < iterationLimit; iteration++){
-            
             initializeLearning(); //set weights to zero
             
             for (int session = 0; session < inputs.length; session++){
                 runTrainingSession(this, inputs[session], expected[session]);
             }
-            
             applyLearning();
         }
     }
-    
-//    public void train(double[][] inputs, double[]expected, int iterationLimit){
-//        for (int iteration = 0; iteration < iterationLimit; iteration++){
-//            
-//            initializeLearning(); //set weights to zero
-//            
-//            for (int session = 0; session < inputs.length; session++){
-//                runTrainingSession(this, inputs[session], expected);
-//            }
-//            
-//            applyLearning();
-//        }
-//    }
+//</editor-fold>
     
     public static void runTrainingSession(NeuralNetIntf net, double[] inputData, double[] expectedData){
-        preparePerceptionLayerForPulse(net, inputData);
+        setInputLayerData(net, inputData);
         net.pulse();
         calculateErrors(net, expectedData);
         calculateAndAppendTransformation(net);
     }
     
-    private static boolean preparePerceptionLayerForPulse(NeuralNetIntf net, double[] inputData) {
-        if (inputData.length != net.getPerceptionLayer().getCount()){
-            System.out.print("Input data size does not match perception layer size.");
+    private static boolean setInputLayerData(NeuralNetIntf net, double[] inputData) {
+        if (inputData.length != net.getInputLayer().getCount()){
+            System.out.print("Input data size does not match input layer size.");
             return false;
         } else {
-            return net.getPerceptionLayer().setOutputValues(inputData);
-        }     
+            return net.getInputLayer().setOutputValues(inputData);
+        }
     }
-
+    
     private static void calculateErrors(NeuralNetIntf net, double[] expectedData) {
-        Sigmoid sigmoid = new Sigmoid();
+        RangeTransformIntf sigmoid = new Sigmoid();
+        
+        //update the output layer from the "known" (expected) answers...
         net.getOutputLayer().updateErrorsFromExpectedResults(expectedData, sigmoid);
-        net.getHiddenLayer().updateErrorsFromConsumerLayer(net.getOutputLayer(), sigmoid);
+        
+        //...then propagate this learning through the model
+        for (int i = net.getLayers().size() - 2; i > 0; i--){
+            net.getLayers().get(i).updateErrorsFromConsumerLayer(net.getLayers().get(i + 1), sigmoid);
+        }
     }
-
+    
     private static void calculateAndAppendTransformation(NeuralNetIntf net) {
-        net.getOutputLayer().updateDeltasBasedOnCurrentErrors(net.getHiddenLayer());
-        net.getHiddenLayer().updateDeltasBasedOnCurrentErrors(net.getPerceptionLayer());
+        for (int i = net.getLayers().size() - 1; i > 0; i--){
+            net.getLayers().get(i).updateDeltasBasedOnCurrentErrors(net.getLayers().get(i + 1));
+        }
     }
-
-    @Override
-    public ArrayList<NeuralLayerIntf> getLayers() {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-
-    @Override
-    public NeuralLayerIntf getLayer(int index) {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-
+//</editor-fold>
+    
 }
